Character Knock Back
Well it works. Explosions now knock the character back. But there are 2
problems with it. The first is that it's weirdly buggy. I think it can
be ironed out through tweaking some variables. The second is that the
whole knockback sequence is instant. There must be a way to move it over
time.