Fixed Rocket Jumping, tweaked mechanics & locked cursor
I fixed rocket jumping. It works pretty damn good now. It did cause
checkpoints to break, but I fixed that. I fixed a lot of the weird bugs
caused by looking down and shooting as well. I tweaked a whole lot of
variables to make the game both challenging, yet achievable. I added in
a cursor lock, so when you enter the game it locks the cursor, allowing
you to play the game without tabbing yourself out. Press ESC to toggle
cursor lock. I upped the sensitivity and increased the FOV to allow
people to make the big movements necessary to play easier, as well as
allow them to see the map better. When I was playing, the vines didn't
work, but in the process of fixing the thing, I changed the type of
character controller being used, so action.cs is a little outdated.