Vortex Rocket and Fire Delay
Added in a fire mode tied to Alpha3 that pulls the player into the
collision origin. Radius and power is adjustable. It needs a HUD button,
though.

Also added a fire delay of 0.5 seconds. This prevents the spamming of
rockets.

Everything should work and build properly. Let me know if you have any
issues. Also, Sung won't read this and complain to me about it later.